// src/app/model/UserWeightObjectifs.ts
export class UserWeightObjectifs {
   id: number;
   typeObjectif: number;
   typeDescription: string;
   percentage: number;
   weight: number;
   reward: string;
   date: string;
   save: boolean;
   edit: boolean;
   openEdit: boolean;
   openConfirmDelete: boolean;
   showNumber: boolean;
 
   constructor() {
      //Constructor initialization 
   }
}